bianace_btcethbnb_trade/requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
requests==2.31.0
python-docx==0.8.11
APScheduler==3.10.4